Run compilation task in an isolated classloader
