+++ gradle-plugin/src/main/kotlin/app/cash/better/dynamic/features/codegen/TypesafeImplementationsCompilationTask.kt
@@ -27,16 +27,9 @@ import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
-import org.jetbrains.kotlin.cli.common.ExitCode
-import org.jetbrains.kotlin.cli.common.arguments.K2JVMCompilerArguments
-import org.jetbrains.kotlin.cli.common.messages.MessageRenderer
-import org.jetbrains.kotlin.cli.common.messages.PrintingMessageCollector
-import org.jetbrains.kotlin.cli.jvm.K2JVMCompiler
-import org.jetbrains.kotlin.config.Services
-import java.io.File
-import java.util.jar.JarEntry
-import java.util.jar.JarFile
-import java.util.jar.JarOutputStream
+import org.gradle.workers.WorkParameters
+import org.gradle.workers.WorkerExecutor
+import javax.inject.Inject
 
 abstract class TypesafeImplementationsCompilationTask : DefaultTask() {
   @get:InputFiles
@@ -54,86 +47,31 @@ abstract class TypesafeImplementationsCompilationTask : DefaultTask() {
   @get:Classpath
   abstract val kotlinCompileClasspath: ConfigurableFileCollection
 
+  @get:Inject
+  abstract val workerExecutor: WorkerExecutor
+
+  interface TypesafeImplementationsCompilationWorkParameters : WorkParameters {
+    val projectJars: ListProperty<RegularFile>
+    val projectClasses: ListProperty<Directory>
+    val output: RegularFileProperty
+    val generatedSources: DirectoryProperty
+    val kotlinCompileClasspath: ConfigurableFileCollection
+    val temporaryDir: DirectoryProperty
+  }
+
   @TaskAction
   fun processImplementations() {
     val tempClassDirectory = temporaryDir.resolve("classes").also { it.mkdirs() }
 
-    compile(tempClassDirectory)
-    mergeClasses(tempClassDirectory)
-  }
+    val workQueue = workerExecutor.classLoaderIsolation()
 
-  private fun compile(classOutputDirectory: File) {
-    val combinedClasspath = (
-      projectClasses.get().map { it.asFile } +
-        kotlinCompileClasspath.files.flatMap { it.walk().filter(File::isFile) }
-      )
-
-    val generatedSourceFiles = generatedSources.asFile.get()
-      .walk()
-      .filter(File::isFile)
-      .map { it.absolutePath }
-      .toList()
-
-    val arguments = K2JVMCompilerArguments().apply {
-      classpath = combinedClasspath.joinToString(separator = File.pathSeparator) { it.absolutePath }
-      destination = classOutputDirectory.absolutePath
-      freeArgs = generatedSourceFiles
-
-      // We supply the stdlib JAR(s) via the classpath, this just suppresses a warning message
-      noStdlib = true
+    workQueue.submit(CompileTypesafeImplementations::class.java) { parameters ->
+      parameters.projectJars.set(projectJars)
+      parameters.projectClasses.set(projectClasses)
+      parameters.output.set(output)
+      parameters.generatedSources.set(generatedSources)
+      parameters.kotlinCompileClasspath.setFrom(kotlinCompileClasspath)
+      parameters.temporaryDir.set(tempClassDirectory)
     }
-
-    // We can't use @SkipWhenEmpty on the sources directory because AGP always needs an output from
-    // this task, but if we try to pass zero source files to the compiler it crashes
-    if (generatedSourceFiles.isNotEmpty()) {
-      val result = K2JVMCompiler().exec(
-        messageCollector = PrintingMessageCollector(
-          System.err,
-          MessageRenderer.GRADLE_STYLE,
-          false,
-        ),
-        services = Services.EMPTY,
-        arguments,
-      )
-
-      require(result == ExitCode.OK) { "Kotlin Compiler Error while compiling dynamic feature implementations." }
-    } else {
-      logger.debug("No generated sources to compile. Skipping.")
-    }
-  }
-
-  private fun mergeClasses(compiledGeneratedClasses: File) {
-    val target = output.asFile.get()
-
-    JarOutputStream(target.outputStream()).use { jar ->
-      // Copy existing jar entries over
-      projectJars.get().map { JarFile(it.asFile) }.forEach { inputJar ->
-        inputJar.entries().iterator().forEach { jarEntry ->
-          jar.putNextEntry(JarEntry(jarEntry.name))
-          inputJar.getInputStream(jarEntry).use { it.copyTo(jar) }
-          jar.closeEntry()
-        }
-      }
-
-      // Copy existing classes over
-      projectClasses.get().map { it.asFile }.forEach { root ->
-        root.walk().filter(File::isFile).forEach { file ->
-          val relative = file.relativeTo(root)
-          jar.putEntry(relative.path, file)
-        }
-      }
-
-      // Copy compiled generated classes over
-      compiledGeneratedClasses.walk().filter(File::isFile).forEach { file ->
-        val relative = file.relativeTo(compiledGeneratedClasses)
-        jar.putEntry(relative.path, file)
-      }
-    }
-  }
-
-  private fun JarOutputStream.putEntry(name: String, source: File) {
-    putNextEntry(JarEntry(name))
-    source.inputStream().copyTo(this)
-    closeEntry()
   }
 }
